requirements.txt
# requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
discord.py>=2.0.0
asyncio-throttle>=1.0.2
